Visualizacion de las tablas en html

+++ src/main/java/com/udea/proyecto/servicios/IServicioEmpresa.java
@@ -1,6 +1,5 @@
 package com.udea.proyecto.servicios;
 
-import com.udea.proyecto.entidades.Empleado;
 import com.udea.proyecto.entidades.Empresa;
 
 import java.util.List;
@@ -11,5 +10,5 @@ public interface IServicioEmpresa {
     public Empresa guardarEmpresa(Empresa empresa);
     public Empresa buscarEmpresaId(Integer numid);
     public Empresa actualizarEmpresa(Empresa empresa);
-    public void borrarEmpresa(Integer numid);
+    public void borrarEmpresa(Long numid);
 }
